requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
python-dotenv>=0.21.0
ollama>=0.1.0
httpx>=0.24.0
